event: common/model, watcher/processor
