+++ nexus-common/src/models/events.rs
@@ -1,58 +1,92 @@
 use crate::db::RedisOps;
 use crate::types::DynError;
-use pubky_app_specs::PubkyId;
+use pubky_app_specs::{ParsedUri, Resource};
 use serde::{Deserialize, Serialize};
+use std::{fmt, path::PathBuf};
+use thiserror::Error;
 use utoipa::ToSchema;
 
-/// Represents a event_list with its public key, URL, and cursor.
-#[derive(Serialize, Deserialize, ToSchema, Default, Debug)]
-pub struct EventsList {
-    pub events: Vec<String>,
-    pub cursor: String,
-    pub limit: usize,
+#[derive(Error, Debug, Clone, Serialize, Deserialize)]
+pub enum EventError {
+    /// The event could not be parsed from a line
+    #[error("InvalidEventLine: {message}")]
+    InvalidEventLine { message: String },
 }
 
-impl RedisOps for EventsList {}
+impl RedisOps for Event {}
 
-impl EventsList {
-    pub async fn new(id: PubkyId) -> Result<Self, Box<dyn std::error::Error + Sync + Send>> {
-        let hs = EventsList {
-            events: vec![],
-            limit: 0,
-            cursor: "0000000000000".to_string(),
+// Look for the end pattern after the start index, or use the end of the string if not found
+#[derive(Debug, Clone, Serialize, Deserialize, PartialEq, ToSchema)]
+pub enum EventType {
+    Put,
+    Del,
+}
+
+impl fmt::Display for EventType {
+    fn fmt(&self, f: &mut fmt::Formatter<'_>) -> fmt::Result {
+        let upper_case_str = match self {
+            EventType::Put => "PUT",
+            EventType::Del => "DEL",
         };
-        // Store event_list with initial cursor in Index
-        hs.put_to_index().await?;
-        Ok(hs)
+        write!(f, "{upper_case_str}")
     }
+}
 
-    /// Retrieves the homeserver from Redis.
-    pub async fn get_from_index(id: &str) -> Result<Option<Self>, DynError> {
-        if let Some(homeserver) = Self::try_from_index_json(&[id], None).await? {
-            return Ok(Some(homeserver));
-        }
-        Ok(None)
-    }
+#[derive(Debug, Clone, Serialize, Deserialize)]
+pub struct Event {
+    pub uri: String,
+    pub event_type: EventType,
+    pub parsed_uri: ParsedUri,
+    pub files_path: PathBuf,
+}
 
-    /// Stores the event_list in Redis.
-    pub async fn put_to_index(&self) -> Result<(), DynError> {
-        self.put_index_json(&[&"foo"], None, None).await?;
-        Ok(())
-    }
-
-    /// Reads from Redis db
-    // TODO: (524)
-    // By Chirs:
-    // > The implementation consist of a new Redis SortedSet where we write down timestamp and PUT/DEL pubky://{pubky}/object (to be crockford32 encoded as in the example above).
-
-    pub async fn from_config(event_list: PubkyId) -> Result<EventsList, DynError> {
-        // Attempt to load the event_list cursor from Redis
-        match EventsList::get_from_index(&event_list).await? {
-            Some(hs) => Ok(hs),
-            None => {
-                // Create a new EventsList instance with default cursor
-                EventsList::new(event_list).await
+impl Event {
+    // TODO: files_path is not used anywhere not sure why it is needed
+    pub fn parse_event(line: &str, files_path: PathBuf) -> Result<Option<Self>, DynError> {
+        let parts: Vec<&str> = line.split(' ').collect();
+        if parts.len() != 2 {
+            return Err(EventError::InvalidEventLine {
+                message: format!("Malformed event line, {line}"),
             }
+            .into());
         }
+
+        let event_type = match parts[0] {
+            "PUT" => EventType::Put,
+            "DEL" => EventType::Del,
+            other => {
+                return Err(EventError::InvalidEventLine {
+                    message: format!("Unknown event type: {other}"),
+                }
+                .into())
+            }
+        };
+
+        // Validate and parse the URI using pubky-app-specs
+        let uri = parts[1].to_string();
+        let parsed_uri =
+            ParsedUri::try_from(uri.as_str()).map_err(|e| EventError::InvalidEventLine {
+                message: format!("Cannot parse event URI: {e}"),
+            })?;
+
+        match parsed_uri.resource {
+            // Unknown resource
+            Resource::Unknown => {
+                return Err(EventError::InvalidEventLine {
+                    message: format!("Unknown resource in URI: {uri}"),
+                }
+                .into())
+            }
+            // Known resources not handled by Nexus
+            Resource::LastRead | Resource::Feed(_) | Resource::Blob(_) => return Ok(None),
+            _ => (),
+        };
+
+        Ok(Some(Event {
+            uri,
+            event_type,
+            parsed_uri,
+            files_path,
+        }))
     }
 }
