+++ nexus-watcher/src/events/processor.rs
@@ -1,14 +1,16 @@
+use super::handlers;
 use super::moderation::Moderation;
-use super::Event;
 use crate::events::errors::EventProcessorError;
 use crate::events::retry::event::RetryEvent;
 use nexus_common::db::PubkyClient;
+use nexus_common::models::events::{Event, EventType};
 use nexus_common::models::homeserver::Homeserver;
 use nexus_common::types::DynError;
 use nexus_common::{get_files_dir_test_pathbuf, WatcherConfig};
 use opentelemetry::trace::{FutureExt, Span, TraceContextExt, Tracer};
 use opentelemetry::{global, Context, KeyValue};
 use pubky_app_specs::PubkyId;
+use pubky_app_specs::{PubkyAppObject, Resource};
 use std::error::Error;
 use std::path::PathBuf;
 use tokio::sync::watch::Receiver;
@@ -173,6 +175,7 @@ impl EventProcessor {
                     info!("Cursor for the next request: {}", cursor);
                 }
             } else {
+                debug!("New event: {}", line);
                 let event = match Event::parse_event(line, self.files_path.clone()) {
                     Ok(event) => event,
                     Err(e) => {
@@ -207,7 +210,7 @@ impl EventProcessor {
     /// # Parameters:
     /// - `event`: The event to be processed
     async fn handle_event(&mut self, event: Event) -> Result<(), DynError> {
-        if let Err(e) = event.clone().handle(&self.moderation).await {
+        if let Err(e) = self.handle(&event).await {
             if let Some((index_key, retry_event)) = extract_retry_event_info(&event, e) {
                 error!("{}, {}", retry_event.error_type, index_key);
                 if let Err(err) = retry_event.put_to_index(index_key).await {
@@ -220,6 +223,113 @@ impl EventProcessor {
         // - ordered set, so that can be ordered by cursor
         // - cursor is local timestamp in milliseconds encoded using crockford32
         // <cursor>: { <PUT|DEL>, <URL> }
+        Ok(())
+    }
+
+    pub async fn handle(&self, event: &Event) -> Result<(), DynError> {
+        // TODO: (524): add method to insert event in to reddis sorted set
+        match event.event_type {
+            EventType::Put => self.handle_put_event(event).await,
+            EventType::Del => self.handle_del_event(event).await,
+        }
+    }
+
+    /// Handles a PUT event by fetching the blob from the homeserver
+    /// and using the importer to convert it to a PubkyAppObject.
+    pub async fn handle_put_event(&self, event: &Event) -> Result<(), DynError> {
+        debug!("Handling PUT event for URI: {}", event.uri);
+
+        let response;
+        {
+            let pubky_client =
+                PubkyClient::get().map_err(|e| EventProcessorError::PubkyClientError {
+                    message: e.to_string(),
+                })?;
+
+            response = match pubky_client.get(&event.uri).send().await {
+                Ok(response) => response,
+                Err(e) => {
+                    return Err(EventProcessorError::PubkyClientError {
+                        message: format!("{e}"),
+                    }
+                    .into())
+                }
+            };
+        } // drop the pubky_client lock
+
+        let blob = response.bytes().await?;
+        let resource = event.parsed_uri.resource.clone();
+
+        // Use the new importer from pubky-app-specs
+        let pubky_object = PubkyAppObject::from_resource(&resource, &blob).map_err(|e| {
+            EventProcessorError::PubkyClientError {
+                message: format!(
+                    "The importer could not create PubkyAppObject from Uri and Blob: {e}"
+                ),
+            }
+        })?;
+
+        let user_id = event.parsed_uri.user_id.clone();
+        match (pubky_object, resource) {
+            (PubkyAppObject::User(user), Resource::User) => {
+                handlers::user::sync_put(user, user_id).await?
+            }
+            (PubkyAppObject::Post(post), Resource::Post(post_id)) => {
+                handlers::post::sync_put(post, user_id, post_id).await?
+            }
+            (PubkyAppObject::Follow(_follow), Resource::Follow(followee_id)) => {
+                handlers::follow::sync_put(user_id, followee_id).await?
+            }
+            (PubkyAppObject::Mute(_mute), Resource::Mute(muted_id)) => {
+                handlers::mute::sync_put(user_id, muted_id).await?
+            }
+            (PubkyAppObject::Bookmark(bookmark), Resource::Bookmark(bookmark_id)) => {
+                handlers::bookmark::sync_put(user_id, bookmark, bookmark_id).await?
+            }
+            (PubkyAppObject::Tag(tag), Resource::Tag(tag_id)) => {
+                if self.moderation.should_delete(&tag, user_id.clone()).await {
+                    Moderation::apply_moderation(tag, event.files_path.clone()).await?
+                } else {
+                    handlers::tag::sync_put(tag, user_id, tag_id).await?
+                }
+            }
+            (PubkyAppObject::File(file), Resource::File(file_id)) => {
+                handlers::file::sync_put(
+                    file,
+                    event.uri.clone(),
+                    user_id,
+                    file_id,
+                    self.files_path.clone(),
+                )
+                .await?
+            }
+            other => {
+                debug!("Event type not handled, Resource: {:?}", other);
+            }
+        }
+        Ok(())
+    }
+
+    pub async fn handle_del_event(&self, event: &Event) -> Result<(), DynError> {
+        debug!("Handling DEL event for URI: {}", event.uri);
+
+        let user_id = event.parsed_uri.user_id.clone();
+        match event.parsed_uri.resource.clone() {
+            Resource::User => handlers::user::del(user_id).await?,
+            Resource::Post(post_id) => handlers::post::del(user_id, post_id).await?,
+            Resource::Follow(followee_id) => handlers::follow::del(user_id, followee_id).await?,
+            Resource::Mute(muted_id) => handlers::mute::del(user_id, muted_id).await?,
+            Resource::Bookmark(bookmark_id) => {
+                handlers::bookmark::del(user_id, bookmark_id).await?
+            }
+            Resource::Tag(tag_id) => handlers::tag::del(user_id, tag_id).await?,
+            Resource::File(file_id) => {
+                handlers::file::del(&user_id, file_id, event.files_path.clone()).await?
+            }
+            other => {
+                debug!("DEL event type not handled for resource: {:?}", other);
+            }
+        }
         Ok(())
     }
 }
